Clients: simplify signal handling
The handler removal utility function previously returned a list of
`asyncio.Handle` objects, which were unusable because they lack any
reference to the original signal obj/name/number. This has been changed
to a list of signal/callback pairs, which can be reused as input for the
add-handler helper utility.